+++ src/GridCal/Engine/Simulations/result_types.py
@@ -51,6 +51,7 @@ class ResultTypes(Enum):
     BusVoltageModule = 'Bus voltage module', DeviceType.BusDevice
     BusVoltageAngle = 'Bus voltage angle', DeviceType.BusDevice
     BusPower = 'Bus power', DeviceType.BusDevice
+    ShadowPrices = 'Bus shadow prices', DeviceType.BusDevice
     BranchOverloads = 'Branch overloads', DeviceType.BranchDevice
     LoadShedding = 'Load shedding', DeviceType.LoadDevice
     ControlledGeneratorShedding = 'Controlled generator shedding', DeviceType.GeneratorDevice